requirements.txt
beautifulsoup4==4.4.1
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
wheel==0.24.0
